Lower profiling array size, fix compîlation warnings
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@3418 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ common/JackNetTool.cpp
@@ -421,14 +421,14 @@ namespace Jack
         dst_params->fPosition.beat = htonl(src_params->fPosition.beat);
         dst_params->fPosition.tick = htonl(src_params->fPosition.tick);
         dst_params->fPosition.bar_start_tick = htonll((uint64_t)src_params->fPosition.bar_start_tick);
-        dst_params->fPosition.beats_per_bar = htonl(src_params->fPosition.beats_per_bar);
-        dst_params->fPosition.beat_type = htonl(src_params->fPosition.beat_type);
+        dst_params->fPosition.beats_per_bar = htonl((uint32_t)src_params->fPosition.beats_per_bar);
+        dst_params->fPosition.beat_type = htonl((uint32_t)src_params->fPosition.beat_type);
         dst_params->fPosition.ticks_per_beat = htonll((uint64_t)src_params->fPosition.ticks_per_beat);
         dst_params->fPosition.beats_per_minute = htonll((uint64_t)src_params->fPosition.beats_per_minute);
         dst_params->fPosition.frame_time = htonll((uint64_t)src_params->fPosition.frame_time);
         dst_params->fPosition.next_time = htonll((uint64_t)src_params->fPosition.next_time);
         dst_params->fPosition.bbt_offset = htonl(src_params->fPosition.bbt_offset);
-        dst_params->fPosition.audio_frames_per_video_frame = htonl(src_params->fPosition.audio_frames_per_video_frame);
+        dst_params->fPosition.audio_frames_per_video_frame = htonl((uint32_t)src_params->fPosition.audio_frames_per_video_frame);
         dst_params->fPosition.video_offset = htonl(src_params->fPosition.video_offset);
         dst_params->fPosition.unique_2 = htonll(src_params->fPosition.unique_2);
     }
@@ -446,14 +446,14 @@ namespace Jack
         dst_params->fPosition.beat = ntohl(src_params->fPosition.beat);
         dst_params->fPosition.tick = ntohl(src_params->fPosition.tick);
         dst_params->fPosition.bar_start_tick = ntohll((uint64_t)src_params->fPosition.bar_start_tick);
-        dst_params->fPosition.beats_per_bar = ntohl(src_params->fPosition.beats_per_bar);
-        dst_params->fPosition.beat_type = ntohl(src_params->fPosition.beat_type);
+        dst_params->fPosition.beats_per_bar = ntohl((uint32_t)src_params->fPosition.beats_per_bar);
+        dst_params->fPosition.beat_type = ntohl((uint32_t)src_params->fPosition.beat_type);
         dst_params->fPosition.ticks_per_beat = ntohll((uint64_t)src_params->fPosition.ticks_per_beat);
         dst_params->fPosition.beats_per_minute = ntohll((uint64_t)src_params->fPosition.beats_per_minute);
         dst_params->fPosition.frame_time = ntohll((uint64_t)src_params->fPosition.frame_time);
         dst_params->fPosition.next_time = ntohll((uint64_t)src_params->fPosition.next_time);
         dst_params->fPosition.bbt_offset = ntohl(src_params->fPosition.bbt_offset);
-        dst_params->fPosition.audio_frames_per_video_frame = ntohl(src_params->fPosition.audio_frames_per_video_frame);
+        dst_params->fPosition.audio_frames_per_video_frame = ntohl((uint32_t)src_params->fPosition.audio_frames_per_video_frame);
         dst_params->fPosition.video_offset = ntohl(src_params->fPosition.video_offset);
         dst_params->fPosition.unique_2 = ntohll(src_params->fPosition.unique_2);
     }
